Rename files and components
Rename "PaymentGameWin" to "Admin"

Previously this project was part of the jacehalvorson/portfolio repo
within a "playoff_bracket" directory. Now that they are in their own
repository, there is no need to have "playoff_bracket" in each of
the names.

File names, component names, and imports are updated in this commit.